+++ app/src/main/java/rsswidget/restwl/com/rsswidget/service/RssDownloadService.java
@@ -1,25 +1,24 @@
 package rsswidget.restwl.com.rsswidget.service;
 
-
 import android.app.Notification;
 import android.app.NotificationChannel;
 import android.app.NotificationManager;
 import android.app.Service;
 import android.content.Context;
 import android.content.Intent;
-import android.content.SharedPreferences;
 import android.os.Build;
 import android.os.IBinder;
 import android.support.annotation.Nullable;
 import android.support.v4.app.NotificationCompat;
+import android.text.TextUtils;
 import android.util.Log;
 
-import java.io.IOException;
 import java.util.List;
 import java.util.concurrent.Executor;
 import java.util.concurrent.Executors;
 
 import rsswidget.restwl.com.rsswidget.R;
+import rsswidget.restwl.com.rsswidget.utils.PreferencesManager;
 import rsswidget.restwl.com.rsswidget.widgedprovider.RssWidgetProvider;
 import rsswidget.restwl.com.rsswidget.network.parsers.XmlParser;
 import rsswidget.restwl.com.rsswidget.database.DatabaseManager;
@@ -27,9 +26,8 @@ import rsswidget.restwl.com.rsswidget.model.RemoteNews;
 import rsswidget.restwl.com.rsswidget.network.HttpConnector;
 
 import static android.support.v4.app.NotificationCompat.PRIORITY_MIN;
-import static rsswidget.restwl.com.rsswidget.utils.Constans.NEWS_COUNT;
-import static rsswidget.restwl.com.rsswidget.utils.Constans.PREFERENCES_KEY;
-import static rsswidget.restwl.com.rsswidget.utils.Constans.TAG;
+import static rsswidget.restwl.com.rsswidget.utils.Constants.ACTION_DATASET_CHANGED;
+import static rsswidget.restwl.com.rsswidget.utils.Constants.TAG;
 
 public class RssDownloadService extends Service {
 
@@ -64,25 +62,24 @@ public class RssDownloadService extends Service {
     public int onStartCommand(Intent intent, int flags, int startId) {
         Runnable runnable = () -> {
             try (DatabaseManager databaseManager = new DatabaseManager(getApplicationContext())) {
-                HttpConnector connector = new HttpConnector("https://lenta.ru/rss/news");
+                String urlString = PreferencesManager.extractUrl(getApplicationContext());
+                if (TextUtils.isEmpty(urlString)) {
+                    stopSelf();
+                    return;
+                }
+                HttpConnector connector = new HttpConnector(urlString);
                 List<RemoteNews> newsList = XmlParser.parseRssData(connector.getContentStream());
                 databaseManager.deleteAllEntryFromNewsTable();
                 databaseManager.insertListNews(newsList);
-                saveInPreferencesNewsCount(newsList.size());
-                RssWidgetProvider.updateSelfData(getApplicationContext());
+                RssWidgetProvider.sendActionToAllWidgets(getApplicationContext(),ACTION_DATASET_CHANGED);
                 Log.d(TAG, "onStartCommand: News download and inserted");
-            } catch (IOException ex) {
+            } catch (Exception ex) {
                 ex.printStackTrace();
             }
             stopSelf();
         };
         executor.execute(runnable);
         return super.onStartCommand(intent, flags, startId);
-    }
-
-    private void saveInPreferencesNewsCount(int count) {
-        SharedPreferences preference = getApplicationContext().getSharedPreferences(PREFERENCES_KEY, Context.MODE_PRIVATE);
-        preference.edit().putInt(NEWS_COUNT, count).apply();
     }
 
     @Override
